third_party/blink/web_tests/external/wpt/css/css-logical/animation-002-expected.txt

This is a testharness.js-based test.
PASS Logical properties can be animated
PASS Logical properties in animations respect the writing-mode
PASS Logical properties in animations respect the direction
PASS Declaration order is respected within @keyframes declaration blocks
PASS Logical properties are able to override physical properties in @keyframes declaration blocks
PASS Declaration order is respected amongst logical properties within @keyframes declaration blocks
PASS Physical properties and logical properties can be mixed
PASS Declaration order is respected on each keyframe individually
FAIL Animations update when the writing-mode is changed assert_equals: expected "50px" but got "0px"
FAIL Filling animations update when the writing-mode is changed assert_equals: expected "100px" but got "0px"
FAIL The number of interpolating properties can be increased when the writing-mode is changed assert_equals: expected "50px" but got "0px"
FAIL The number of interpolating properties can be decreased when the writing-mode is changed assert_equals: expected "150px" but got "200px"
FAIL Animations update when the writing-mode is changed through a CSS variable assert_equals: expected "50px" but got "0px"
PASS Logical shorthand with variable references animates correctly
FAIL Animations update when the direction is changed assert_equals: expected "0px" but got "50px"
Harness: the test ran to completion.
